Ai_attendence_system/server/app/requirements.txt
anyio==4.9.0
asgiref==3.8.1
attrs==24.3.0
channels==4.2.2
click==8.1.8
Django==5.1.4
django-cors-headers==4.6.0
django-extensions==4.1
djangorestframework==3.15.2
dlib==19.24.8
drf-spectacular==0.28.0
face-recognition==1.3.0
face_recognition_models==0.3.0
geographiclib==2.0
geopy==2.4.1
h11==0.16.0
httptools==0.6.4
idna==3.10
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
numpy==2.2.1
packaging==24.2
pandas==2.2.3
pillow==11.2.1
psycopg2==2.9.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
rpds-py==0.22.3
setuptools==75.1.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2024.2
uritemplate==4.1.1
uvicorn==0.34.2
uvloop==0.21.0
watchfiles==1.0.5
websockets==15.0.1
wheel==0.44.0
